+++ CHANGELOG.rst
@@ -6,8 +6,9 @@ Current
 
 - **Breaking**: the *BASIC* `ACL` engine will now grant users on all agents if they are not explicitly defined
 - Add: new plugins system to allow users to write their own modules
-- Add: `support new burp counters <https://git.ziirish.me/ziirish/burp-ui/issues/219>`_
+- Add: `backups deletion <https://git.ziirish.me/ziirish/burp-ui/issues/203>`_
 - Add: `record login failure attempt <https://git.ziirish.me/ziirish/burp-ui/issues/214>`_
+- Add: `support new burp counters <https://git.ziirish.me/ziirish/burp-ui/issues/219>`_
 
 0.5.1 (05/26/2017)
 ------------------